Docker/chat_robot/requirements.txt
tornado==6.0.2
tencentcloud-sdk-python==3.0.94